Use FxHash everywhere
