+++ clippy.toml
@@ -1,1 +1,3 @@
-cognitive-complexity-threshold = 20
+disallowed-types = [
+    { path = "std::collections::HashMap", reason = "Use utils::hasher::HashMap" },
+]
